+++ stwo_cairo_prover/crates/cairo-air/src/verifier.rs
@@ -18,12 +18,12 @@ use stwo_constraint_framework::PREPROCESSED_TRACE_IDX;
 use thiserror::Error;
 use tracing::{span, Level};
 
-use crate::air::{
-    lookup_sum, CairoClaim, CairoComponents, CairoInteractionElements, MemorySection, PublicData,
-    PublicMemory, PublicSegmentRanges, SegmentRange,
-};
+use crate::air::{lookup_sum, CairoClaim, CairoComponents, CairoInteractionElements};
 use crate::builtins_air::BuiltinsClaim;
 use crate::components::memory_address_to_id::MEMORY_ADDRESS_TO_ID_SPLIT;
+use crate::public_data::{
+    MemorySection, PublicData, PublicMemory, PublicSegmentRanges, SegmentRange,
+};
 use crate::{CairoProof, PreProcessedTraceVariant};
 
 fn verify_claim(claim: &CairoClaim) {
